move xcm weight misc to primitives
