+++ pallets/liquid-staking/src/types.rs
@@ -1,6 +1,5 @@
 use codec::{Decode, Encode};
 
-use frame_support::weights::Weight;
 use scale_info::TypeInfo;
 use sp_runtime::{
     traits::{AtLeast32BitUnsigned, Zero},
@@ -55,36 +54,5 @@ impl<Balance: AtLeast32BitUnsigned + Copy + Clone> MatchingLedger<Balance> {
     #[inline]
     pub fn is_matched(&self) -> bool {
         self.total_stake_amount == self.total_unstake_amount
-    }
-}
-
-/// The xcm weight when execute staking call wrapped in xcm message
-#[derive(Copy, Clone, Eq, PartialEq, Encode, Decode, RuntimeDebug, TypeInfo)]
-pub struct XcmWeightMisc<Weight> {
-    /// The weight when execute bond xcm message
-    pub bond_weight: Weight,
-    /// The weight when execute bond_extra xcm message
-    pub bond_extra_weight: Weight,
-    /// The weight when execute unbond xcm message
-    pub unbond_weight: Weight,
-    /// The weight when execute rebond xcm message
-    pub rebond_weight: Weight,
-    /// The weight when execute withdraw_unbonded xcm message
-    pub withdraw_unbonded_weight: Weight,
-    /// The weight when execute nominate xcm message
-    pub nominate_weight: Weight,
-}
-
-impl Default for XcmWeightMisc<Weight> {
-    fn default() -> Self {
-        let default_weight = 3_000_000_000;
-        XcmWeightMisc {
-            bond_weight: default_weight,
-            bond_extra_weight: default_weight,
-            unbond_weight: default_weight,
-            rebond_weight: default_weight,
-            withdraw_unbonded_weight: default_weight,
-            nominate_weight: default_weight,
-        }
     }
 }
